pyproject.toml
[tool.poetry]
name = "gitcoin-search-spike"
version = "0.1.0"
description = ""
authors = ["Massimiliano Mirra"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
langchain = "^0.0.279"
python-dotenv = "^1.0.0"
openai = "^0.28.0"
tiktoken = "^0.4.0"
rich = "^13.5.2"
jq = "^1.6.0"
sentence-transformers = "^2.2.2"
torch = ">=2.0.0, !=2.0.1"
rake-nltk = "^1.0.6"
fastapi = "^0.103.2"
markdown = "^3.4.4"
strip-markdown = "^1.3"
mdx-linkify = "^2.1"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
chromadb = "^0.4.13"
lunr = "^0.7.0.post1"
pytest-watcher = "^0.3.4"
numpy = "^1.26.1"
apscheduler = "^3.10.4"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}


[tool.poetry.group.dev.dependencies]
taskipy = "^1.12.0"
pytest = "^7.4.1"
syrupy = "^4.5.0"
mypy = "^1.5.1"
chromadb = "^0.4.8"
pytest-only = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
test_watch = "ptw . -- --capture=no -vv"
test = "pytest"
run = "PYTHONPATH=$PWD python src/main.py"

[tool.pyright]
include = ["src", "tests"]
exclude = ["**/__pycache__"]
ignore = []
defineConstant = { DEBUG = true }

[tool.black]

[tool.pytest-watcher]
now = true
